Modified controller to respect single responsibility principle
